apps/Realtime-Image-Transmission/pyproject.toml
[tool.coverage.run]
branch = true
source = [ "realtime_image_transmission" ]

[tool.coverage.report]
exclude_lines = ['if TYPE_CHECKING:']
show_missing = true

[tool.pytest.ini_options]
addopts = "--cov --cov-fail-under=100 --cov-report html:'../../coverage/apps/realtime-image-transmission/html' --cov-report xml:'../../coverage/apps/realtime-image-transmission/coverage.xml' --junitxml='../../reports/apps/realtime-image-transmission/unittests/junit.xml' --html='../../reports/apps/realtime-image-transmission/unittests/html/index.html'"

[tool.poetry]
name = "PyPackages"
version = "1.0.0"
description = ""
authors = [ ]
license = 'Proprietary'
readme = 'README.md'

  [[tool.poetry.packages]]
  include = "realtime_image_transmission"

  [tool.poetry.dependencies]
  python = ">=3.8,<3.11"
  opencv-python = "^4.10.0.84"
  websocket-client = "^1.8.0"
  websockets = "^13.1"
  python-socketio = "^5.11.4"
  requests = "^2.32.3"

  

[build-system]
requires = ["poetry-core==1.1.0"]
build-backend = "poetry.core.masonry.api"
